fix(yaml_to_cxx): nested array initializers must be wrapped in {}
According to the spec at most one user-defined conversion is
allowed in a user-defined conversion sequence (12.3p4). Therefore
when converting the python tests we cannot allow Array{Array{}}, as
the implicit conversion to Datum will not occur. This is (for now)
fixed by wrapping child elements in another set of braces.