refactor(connection): clean up connection socket creation
While its not technically broken, setting up the socket all inside
the constructor could lead to potentially undefined behavior.
Instead all of the socket-related activity is moved to R::connect
and all moved into the Connection if all goes as expected. This is
in line with the asio code, so it serves to make them both look a
little closer for a potential future where they are pluggable
backends